+++ src/librt/globals.c
@@ -1,7 +1,7 @@
 /*                       G L O B A L S . C
  * BRL-CAD
  *
- * Copyright (c) 2004-2010 United States Government as represented by
+ * Copyright (c) 2004-2011 United States Government as represented by
  * the U.S. Army Research Laboratory.
  *
  * This library is free software; you can redistribute it and/or
@@ -281,12 +281,12 @@ fastf_t rt_cline_radius = (fastf_t)-1.0;
 /**
  * minimum number of bot pieces
  */
-int rt_bot_minpieces = RT_DEFAULT_MINPIECES;
+size_t rt_bot_minpieces = RT_DEFAULT_MINPIECES;
 
 /**
  * minimum triangles per piece
  */
-int rt_bot_tri_per_piece = RT_DEFAULT_TRIS_PER_PIECE;
+size_t rt_bot_tri_per_piece = RT_DEFAULT_TRIS_PER_PIECE;
 
 /**
  * rt vlist command descriptions
