+++ src/librt/primitives/hf/hf.c
@@ -1,7 +1,7 @@
 /*                            H F . C
  * BRL-CAD
  *
- * Copyright (c) 1994-2010 United States Government as represented by
+ * Copyright (c) 1994-2011 United States Government as represented by
  * the U.S. Army Research Laboratory.
  *
  * This library is free software; you can redistribute it and/or
@@ -1600,8 +1600,8 @@ rt_hf_plot(struct bu_list *vhead, struct rt_db_internal *ip, const struct rt_tes
     vect_t zbasis;
     point_t start;
     point_t cur;
-    int x;
-    int y;
+    size_t x;
+    size_t y;
     int cmd;
     int step;
     int half_step;
@@ -1734,10 +1734,10 @@ rt_hf_plot(struct bu_list *vhead, struct rt_db_internal *ip, const struct rt_tes
 
     /* Apply relative tolerance, if specified */
     if (!NEAR_ZERO(ttol->rel, SMALL_FASTF)) {
-	int rstep;
+	size_t rstep;
 	rstep = xip->w;
 	V_MAX(rstep, xip->n);
-	step = (int)(ttol->rel * rstep);
+	step = ttol->rel * rstep;
     } else {
 	/* No relative tol specified, limit drawing to 'goal' # of vectors */
 	if (goal <= 0) return 0;		/* no vectors for interior */
