+++ src/libwdb/dsp.c
@@ -1,7 +1,7 @@
 /*                            D S P . C
  * BRL-CAD
  *
- * Copyright (c) 1994-2010 United States Government as represented by
+ * Copyright (c) 1994-2011 United States Government as represented by
  * the U.S. Army Research Laboratory.
  *
  * This library is free software; you can redistribute it and/or
@@ -34,7 +34,7 @@
 #include "db.h"
 
 int
-mk_dsp(struct rt_wdb *fp, const char *name, const char *file, int xdim, int ydim, const matp_t mat)
+mk_dsp(struct rt_wdb *fp, const char *name, const char *file, size_t xdim, size_t ydim, const matp_t mat)
     /* name of file containing elevation data */
     /* X dimension of file (w cells) */
     /* Y dimension of file (n cells) */
