+++ src/sig/imod.c
@@ -1,7 +1,7 @@
 /*                          I M O D . C
  * BRL-CAD
  *
- * Copyright (c) 1986-2010 United States Government as represented by
+ * Copyright (c) 1986-2011 United States Government as represented by
  * the U.S. Army Research Laboratory.
  *
  * This program is free software; you can redistribute it and/or
@@ -186,9 +186,7 @@ int main(int argc, char **argv)
     }
 
     if ( clip_high != 0L || clip_low != 0L ) {
-	(void)fprintf( stderr, "%s: clipped %lu high, %lu low\n",
-		       progname,
-		       clip_high, clip_low );
+	(void)fprintf( stderr, "%s: clipped %lu high, %lu low\n", progname, (long unsigned)clip_high, (long unsigned)clip_low );
     }
 
     return 0;
